requirements.txt
streamlit
azure-ai-projects>=1.1.0b4
azure-identity
streamlit_msal
aiohttp
pyperclip